fix: Merge errors and rename conflicts
